backend-software/requirements.txt
asgiref==3.4.1
astroid==2.8.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-mock-queries==2.1.7
djangorestframework==3.12.4
docker==5.0.2
greenlet==1.1.1
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
model-bakery==1.3.2
numpy==1.21.2
openapi-codec==1.3.2
packaging==21.0
pandas==1.3.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
responses==0.14.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
